Allow lng/lat to be set from imported CSV
This significantly reduces our dependency on external geocoders, which is particularly helpful in testing.

It's a bit ugly because of the way mongoid handles before_save callbacks - they'll be called at only one layer of the documents so we need a callback at each layer depending on how the documents are being created. We can clean this up soon.