Automatically create mongo indexes in test
This means not having to run a rake task after each index change. On the one hand, the manual step is a good reminder that we will have to run that task when deploying to production. On the other it was slowing down development quite a bit. We should probably build the index creation task into deployments.