Refactor places controller to tidy up code. Move some logic to model.
Main aim here was to make the controller methods more succinct. I added some tests (which are really integration tests as they're testing our integration with the database, but can live here for now) and moved some of the logic into a scope on the Place model.

I'm not sure that the calculations to make distances work in the expected units work. I'll be testing that next.