Implement Places.find_near method to find places within a given distance
Mongoid supports Mongo's 'near' criterion but for various reasons it's preferable to use geoNear, chiefly that it will return the distance from each result to the origin. As a result we're talking directly to mongo (via the ruby driver) and then explicitly casting results to mongoid objects.

The distance calculations here aren't correct. At short distances they under-estimate the distance between points and at long distances they significantly over-estimate them. That could be to do with the curvature of the earth and the fact you can't cleanly convert miles to degrees in the way this code attempts?