Add rake task to make associations where none exist.
Where a business support scheme has no relations in a collection (e.g. locations) it is inferred that it belongs to all locations.
This has lead to scheme retrieval becoming more complex than it needs to be. By adding all associations with this task the additional query clauses can eventually be removed.