Refactor the active method.
The number processing way of doing it was pretty rough and not really
obvious. Create two private methods to determine if a scheme is started
or ended, and then active? method is just a combination of those.

has_started? and has_ended? could be made public easily enough but we
don't use them elsewhere so it makes no sense to right now.