Disable spherical coordinates geoNear query option.
MongoDB 2.0 does not support indexes using a spherical coordinate system.
Until now we have not been using the index because it is not compatible with a spherical query.
This changes the geoNear query to use a cartesian distance calculation which is compatible with the existing index.
We had to change the unit tests to take into account the different distance values calculated using this new coordinate system.