Change how sidekiq is configured and started.
This adds a Procfile with an entry for starting sidekiq.  This will
enable us to transition from using the rake task and the delayed_job
worker to using the more generic procfile worker.  This makes it more
consistent with other apps.

This also now sets sidekiq options in a `sidekiq.yml` file.  This is
consistent with how we do it in other apps.  This change also removes
the specification of which queues to listen to.  By default, sidekiq
listens to the queue called 'default', which is what we want.  It
previously also listened to the 'failed' queue, which I think was a
mistake.  We're not using this queue explicitly anywhere, so adding it
seems unnecessary.