Remove redundant geocoding cronjob and script
geocoding has been done by a sidekiq worker for a long time now, so this
ronjob is redundant.  Additionally, it doesn't have a distributed lock
to ensure it will only run on a single server at a time, which could
lead to interesting data corruption.  This script was the only thing
using the lockfile gem, so that's gone as well.

This leaves whenever itself in place so that the existing cronjobs can
be cleaned up.