Upgrade to ruby 2.2.3
We upgrade kgio, rainbows, and eventmachine to get versions that
will compile against ruby 2.x.  We also have to add the test-unit gem in order
to be able to run the tests.  This may introduce confusion as the app
explicitly uses minitest too.  Hopefully this is temporary until we get to
rails 4.x though.