Upgrade webmock
We go to the latest version possible 1.17.4.  Later versions require a newer
minitest which is probably not compatible with this rails.  We also have to
pull in the webmock/test-unit instead of webmock/minitest hooks.  Suspect
this will change if we upgrade to rails 4.x where we can drop the test-unit
dependency entirely.