Bump cucumber
We go up to the last of the 1.x series to get full ruby 2.x support.  It
happily ran the features without this update but the runner would try to
re-invoke itself after running and fail causing the test run to report a
failure even though all the tests passed.