Be more explicit about mapit calls in tests
The geocode process in Place objects has robost error handling to avoid
failing should the mapit communication go wrong: it saves the error on the
object being saved and this allows the object to save without geocode
information.  Unfortunately when we bumped webmock its
`NetConnectNotAllowedError` stopped being an `Error` and started being an
`Exception`.  This means it is no longer trapped by `resuce => e`.  It's
not a good idea to capture `Exception` in your own code as it's the base
class for things like SIG_INT etc.  This means we have to be more explicit
in our tests about setting up webmock expectations for the postcode geocoding
calls we are going to make.  In all cases we don't rely on the Place objects
we create actually having geocode information so we make use of the
`mapit_does_not_have_a_postcode` method from gds-api-adapters to stub out a the
call as a failed lookup.