Stop validating that CSVs are valid HTML
The intention of this was to make sure that no sketchy HTML is present in the
fields of the CSV and so upstream consumers of the places API don't risk
presenting that HTML when they render places.  A noble goal, but poorly
implemented as it only protected uploaded Places, not any entered or edited
directly via the admin interface.

The other problem with the solution as it stands is that it uses govspeak to
validated the file.  Govspeak has a known problem for apps that want to upgrade
to ruby 2.2.x (see: https://github.com/alphagov/govspeak/issues/57) and so this
is a blocker for our upgrade.

Having discussed the issue with @alext and @jennyd we decided that to proceed we
would drop the feature entirely to avoid blocking the upgrade and create a story
in the backlog (see: https://trello.com/c/jFShPn1t/167) to properly address
sanitising or rejecting Place data at a later date.