Bump simplecov
We also need to move the setup for coverage of cucumber out of coverage.rb
and add it to the very top of env.rb which is loaded first.  As per
documentation.  Without this we seem to try and run cucumber a 2nd time with
invalid arguments when trying to generate the coverage report.