Remove shoulda and only use shoulda-context
We weren't using any of the shoulda-matchers, so there's no need to have the
full "shoulda experience".  This also reveals that we can't include
MiniTest::Assertions anymore either - shoulda or shoulda-matchers must have
been doing some massaging of namespaces to make that module available.  The
specs all run fine without this include and it removes the warning output.