Bump capybara and poltergeist
Because new capybara is more strict than old capybara we have to change a few
feature step definitions to be more explicit about what they are interacting
with and this does involve html changes to introduce a class or two. Nothing
too drastic though.  We also configure capybara to use poltergeist as the
javascript driver when run via cucumber - something it wasn't doing previously
(not that we have any JS cucumber features, but you never know).