Setup ci-reporter for jenkins correctly
Now that we have to include testunit and minitest the app seems to think it's
using testunit not minitest so the ci:setup:minitest task doesn't end up
generating the report files.  This means that our build won't pass as a lack of
reports is a considered a failure by ci.

Once we're on rails 4.x we can review again which framework we are using for tests; minitest vs. test-unit and only invoke one of the setup tasks.

Tried to add reporting for cucumber, but no version worked satisfactorially. It
either made cucumber not load any features, and thus report an error or trapped
the cucumber output so the ci console reported none and you had to click
through to the junit reports to see the problems.  This doesn't seem like a great idea as it changes dev workflow and not neccessarily for the better.  So
we left it out.  Perhaps we can revisit later.