Remove extra England region from EUR areas response
Since 63a371af7bdee5b9f08c22d759a743ea8b201618 we have added an `England`
region into the EUR areas response. I think the intent was to have this area
available to be selected for business support schemes in Publisher, to match
previous behaviour when 'locations' were used (before 'areas' were added for
postcode lookup).

However, although this area could be selected in Publisher for a business
support scheme, it didn't function as intended for users of the finder: the
area didn't exist in MapIt so would never be returned for a postcode lookup,
so would never appear in the facets used for querying Content API for schemes.
If England was the only area selected for a scheme, that scheme would never
appear in finder results when filtering by postcode. In practice this
generally wasn't a problem because other regions covering England would also
be selected and the scheme would show up via those areas, but the England
region wasn't adding any value and was misleading to content designers.

Since we moved to using GSS codes instead of slugified names of areas to
find business support schemes by postcode, the England region has no longer
appeared in the available areas to tag by in Publisher because Imminence
wasn't adding in a GSS code for it.

England is not an electoral region for the European Parliament (the EUR type
in MapIt) so it's misleading to include it in a response for areas of this
type. These areas have (and have had) a variety of other uses as well [1].

This commit stops adding the England region into the EUR areas response.

From looking at the last week of data in Kibana, only Publisher uses the EUR
areas endpoint, and I think that's only in relation to business support
schemes:

- getting the options for the areas multiselect on the edition edit page
- getting the area names for the CSV export

Both of these continue to work with this change.

[1]: https://en.wikipedia.org/wiki/Regions_of_England