Include all place fields in dataset CSV export
We've seen instances of data loss in Production in the past because users
export a CSV, edit it and upload it to create a new dataset version; not all
fields were exported so some data was lost. When the service uses local
authority lookup and the SNACs are lost, no results at all will be returned
to users.

This change includes all of the fields from the Place model (except the _id)
in the CSV export, as the JSON format does already. We have to handle a few
fields separately, though:

- location: this is a Point object so can't be serialised directly to CSV. The
  point either comes from the postcode (from MapIt) or from the override
  fields so doesn't need to be included in the export in order to be preserved
  on import.
- override_lng and override_lat: these values do need to be preserved if
  they're present, but they need to have different names in order for that to
  happen :( The import uses Place.parameters_from_hash, which looks for fields
  called 'lng' and 'lat' in the input and returns those as the override
  fields. Unfortunately those names are less specific and so could be
  misleading in the CSV since only a minority of rows will contain values for
  them, implying that most rows don't have locations (because of unreported
  geocode errors?) when in practice they do. It would be better to make
  Place.parameters_from_hash look for the override names in the input, but we
  need to speak to the Content team to find out how these CSVs are used before
  changing that.

The latter point affects the JSON export as well: in the past we have worked
around the lack of all the fields in the CSV by exporting to JSON, converting
that to CSV, editing and importing, but we may have lost override locations by
doing that because the field names didn't match.