Test exporting to CSV and then uploading the CSV
At the moment it's hard to find individual records in an Imminence dataset to
edit it in the app, particularly for the larger datasets with multiple pages
of places. A workaround for this is to export the dataset to CSV, edit it and
re-upload it. We need to be confident that we're not losing data via this
workflow; in the past we've accidentally lost all the SNACs from a dataset by
doing this, because they weren't exported to CSV.

This Cucumber scenario tests this workflow by exporting a dataset to CSV,
re-uploading the CSV, and checking that the place created from the uploaded
data is identical to the pre-existing one (except for its id and the dataset
version number). The pre-existing place has values set for all attributes
(except geocode error - we're overriding the location so there's no need to
geocode from the postcode and we wouldn't want to preserve the error message
on import anyway).