Use new jenkins.sh from govuk-rails-app-template
We noticed that some of our branch builds were marked as "unstable" in jenkins
and were being reported to github as successes.  Leading us to merge things
that weren't actually passing.

The reason is that our jenkins_branches.sh build uses the exit code of
jenkins.sh to decide what to tell github.  Prior to c1a16d9f81f4f4d83d022f40234d8f61aed46b94
our build script was just to run the tests via rake and the exit code of that
rake task was the exit code of the whole script.  After this commit we added a
second step to test compiling the assets, now if the tests failed the script
proceed to compile the assets regardless and use the exit code of that for the
exit code of the whole script.  This gave us false positives.

We solve it by using the new combined jenkins.sh from
govuk-rails-app-template[1] and modifying it slightly following the example of
whitehall[2] to cope with our multiple stage build.  This new script uses a
trap function to capture errors at any stage in the script and tell github
about the failure as well as aborting the script.  We lose the ability to see
all errors in one run (e.g. if we commit something that breaks the tests and
asset compilation we need to run two builds to find this out), but we also lose
the false positives and gain parity with other GOV.UK apps.

[1]: https://github.com/alphagov/govuk-rails-app-template/blob/e7f22d96879f974dffcf5d2b4f61e0509f19069e/templates/jenkins.sh.erb
[2]: https://github.com/alphagov/whitehall/blob/34b19e1118579fbe36463744f0a3f3f1ffcc1c93/jenkins.sh#L63-L75