Use transient factorygirl attributes in Places for lat/long
This allows us to not have to remember to construct a Point for location
whenever we want a place with a different lat/long to the default factory
values.  While we're at it we explode out our calls to FactoryGirl
when building Places so they're attribute per line, and easier to see.