Create CsvData model
This is to hold CSV data uploaded for new data sets.  These will not be
stored embedded in the data set (which is in turn embedded in the
service).  This means we avoid issues with breaching per-document limits
on size if any csv processing job is killed without completing.

We have followed the pattern on Place for storing the connection between
the csv data and the data set it "belongs to".  The reason for doing it
this way rather than using a traditional has_one/belongs_to reference
relationship is that mongoid does not support these when one end of the
association is an embedded document.