Plug CsvData into DataSet
We try to keep the interface the same, with the only external change
being that DataSet no longer has a csv_data setter (instead you always
use the data_file setter) and the csv_data getter now returns a CsvData
not a string.

Internally we have to add some more callbacks to make sure the CsvData
is saved during the normal life-cycle of the DataSet and that any errors
from the CsvData are pushed back onto the DataSet.

Using has_one / belongs_to reference associations would avoid much of this
however reference associations between mongoid documents don't work if one
of the documents is embedded in another.  So we have to hand-roll it.