Make Service#schedule_csv_processing use the DataSet implementation
We have both Service#schedule_csv_processing and
DataSet#schedule_csv_processing because after_save callbacks on embedded
documents are not run when they are saved from within the document they
are embedded in.  Now that the csv_data is an external document we need
to do slightly more work in DataSet#schedule_csv_processing than we used
to.  Rather than trying to duplicate the effort we make the version in
Service call down to the version on DataSet by passing around a reference
to the DataSet we want to do the csv processing on.