Propagate errors from CsvData to DataSet and Service
To get the errors on the csv data to display in the front end it is
easiest if they live on the object rendered in the form (e.g. the data et
or the service).  We use callbacks to propagate any errors on the data
attribute of CsvData to the data_file attribute of the DataSet and then
up to the data_file attribute of the Service if present.