Add feature tests to check for version panels in history tab
We test that the correct number of version panels are showing for a Service as well as that the data shown is correct for the newly created version and dataset. We use CSS nth-child selectors in favour of other approaches (such as adding extra CSS classes or ids to the html tags).