Refactor Service creation through form submit for features
The form for creating a Service has gained more fields such as 'Location match type' and 'Local authority hierarchy match type'. We refactor the step definition to accept a table of parameters to allow for more comprehensive testing.
We provide a hash of default values which get merged before the Service is created so that we do not have to pass all parameters every time when making use of this step.