Opt-in to new gds-api-adapters behaviour for 404, 410, and openstruct
This is made slightly confusing by the behaviour of the mapit api adapter
which doesn't return the response directly, unlike other adapters, and
instead returns a GdsApi::Mapit::Location wrapper object which can expose
some openstruct instances.  We try to update the callsites for where we
interact with those openstructs to use the hash-accessor `[]` instead
of direct method access, to future-proof us against any change of those
openstructs into hashes.