Change Mongoid to not raise DocumentNotFound errors
Rubocop wants to change `.where(...).first` to `.find_by(...)` and in
the case of Mongoid this changes the semantics to error if nothing is
found, rather than evaluating to `nil`.

Therefore, just change the configuration of Mongoid to not raise an
exception in this case. This is more consistent with ActiveRecord,
which might be useful anyway.