Fix linting issues
Previously we auto-generated the cucumber.rake file, but only use
a small part of it. This removes the unnecessary parts that were
causing linting issues for this file.

Previously we defined a task to ensure we used zone-sensitive times,
but this is now provided by RuboCop [1]. This removes the superseded
task that had an associated linting issue.

[1]: https://www.rubydoc.info/gems/rubocop/0.41.2/RuboCop/Cop/Rails/TimeZone