Upgrade Gems Mongo & Mongoid
RE GOVUK are in the process of migrating from self hosted Mongo instances to AWS DocumentDB.
DocumentDB is running the equivalent of MongoDB 3.6.0 (a higher version than we currently have on our server). This means that the following gems need to be updated accordingly:
mongoid to 6.3.0
mongo to  2.5.0

Mongo has now been upgraded to 2.6.12 (from 2.4.9) so this update should not cause
any errors. Please see the Ruby/Mongo driver compatibility chart for more
information:
https://docs.mongodb.com/ruby-driver/master/reference/driver-compatibility/