Lint as part of the default rake task
We had expected that all apps were linting with the default rake task
and removed the linting step from the Jenkins library [1]. However it
seems this app was missed.

To add this to the default rake task we clear the original one to
provide clarity as to what it actually does (which seems to be run test
and cucumber). The Rails test task seems to execute in a strange way
making it the last task so to work with that I just defined cucumber to
come before it.

As test tasks aren't loaded in production environments there is the
check as to whether the task is defined. Otherwise an error occurs if
Rake is clearing a non-existent task.

[1]: https://github.com/alphagov/govuk-jenkinslib/pull/74