Replace inefficient `empty?` conditionals
The use of `empty?` in a number of conditional statements causes
performance issues (the query takes more than 1s to execute for
each dataset, so the page often times out).

Switching to use the count instead, which performs a more efficient
database query (less than 0.001s per query). Instead of loading the full
results (with `empty?`), we are now just getting a count.