Modify tests to include Rspec matchers
Rspec matchers are necessary as Rspec is a dependency of pact, and having Rspec
present changes the implemention of some assertions due to the naming clashes between
Rspec and minitest. Simply changing the expectations of the tests is the easiest way round
this, the only other way would be to change the entire repo to use Rspec, which
would be a large amount of work.