WIP: start updating Minitest assertions to RSpec expectations
The inclusion of Pact means the inclusion of RSpec (one of its
dependencies), whose presence seems to interfere with how
cucumber-rails works. Switching to RSpec style assertions fixes
the issue (at least in data_sets.feature:96) and is just as
readable. It also frees us up to remove Minitest in the future.