Prevent CSV processing race condition
This fixes a race condition where a CSV file would
be uploaded but not replicated to the read-nodes
before processing. This ensures that when the
Sidekiq job fetches the CSV it exists in the db
node (read-replicas). This is achieved by waiting
until the majority of nodes have acknowledged the
write.