Add sprockets-rails gem
The sprockets-rails gem is implicitly included as a dependency of
sassc-rails, but if in future we take sassc-rails out, we'd have
to know to put the sprockets-rails gem dependency back in. Since
we're not 100% sure how that missing dependency would manifest
itself (safely at dev/test time or unsafely later on), safer to
leave it in as an explicit dependency.