Allow a more realistic interpretation of csv mime types
The method we use to detect csv files is pretty shaky - it
returns different values in docker, on the desktop, and
remotely. Unfortunately, there's not really any improvement
on it - csv files are essentially detected by "this isn't
any other type of file, so it must be text I guess". By
opening up the mime types considered as CSV, we allow the
test suite to run on govuk-docker again. It's arguably a
bit of test-induced damage, but I couldn't find a better
solution that wasn't heavy duty.