Update distances in test - these seem to be more accurate(??)
Test values under PostGIS seem radically different from those
in Mongoid, but match far more closely to the measured results
in Google Maps. I suspect this is correct (ie PostGIS is doing
this better than Mongoid was), but I will get other people
to check this.

Note that test queries have to be unscoped before #count() can
be used in them because we have custom selects to get the
distances, and PG select changes interfere with #count().