Call delete Sidekiq worker when activating dataset
Currently Sidekiq is used for archiving places. Following this pattern,
a new Sidekiq worker has been created to delete historic data sets /
place archives if there are any eligible.

We will run the worker when a new data set is activated as this
process generates a new archived data set.